Fix compiler warnings

+++ src/messages_host.c
@@ -5,7 +5,7 @@ int main(int argc, char **argv)
 	bsp_init("messages_kernel.elf", argc, argv);
 	bsp_begin(16);
 
-	size_t tagsize = sizeof(int);
+	int tagsize = sizeof(int);
 	ebsp_set_tagsize(&tagsize);
 
 	ebsp_spmd();
